rag_api_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
supabase # Required by langchain-supabase & direct client use
# LangChain core and integrations
langchain
langchain-ollama
# Removed: langchain-postgres
# Removed: psycopg2-binary 
# Removed: SQLAlchemy
langchain-community # Contains community integrations like Supabase
langchain-openai # Added for OpenAI LLM support 
python-jose[cryptography] # For decoding and validating MSAL JWTs 